requirements.txt
amqp==2.5.2
appdirs==1.4.3
asgiref==3.2.7
billiard==3.6.3.0
cached-property==1.5.1
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-admin-interface==0.12.2
django-admin-notifications==0.6.3
django-colorfield==0.3.0
django-extensions==1.9.8
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-jet==1.0.8
django-kronos==1.0
django-pardakht==1.4.2
idna==2.9
importlib-metadata==1.6.0
isodate==0.6.0
kombu==4.6.8
lxml==4.5.0
mysqlclient==1.4.6
Pillow==7.1.1
python-crontab==2.4.2
python-dateutil==2.8.1
pytz==2019.3
redis==3.5.0
requests==2.23.0
requests-toolbelt==0.9.1
six==1.14.0
sqlparse==0.3.1
typing==3.7.4.1
urllib3==1.25.9
vine==1.3.0
zeep==2.5.0
zipp==3.1.0
